requirements.txt
streamlit==1.35.0         # UI for the interactive dashboard
pandas==2.2.2             # Data manipulation and analysis
nltk==3.8.1               # Natural language processing toolkit
textblob==0.17.1          # Simple sentiment analysis built on top of NLTK
matplotlib==3.7.0         # Visualization for data insights
openai==1.13.3            # OpenAI API for GPT-4 summaries
python-dotenv==1.0.0      # Load API keys from .env files
httpx==0.27.0             # Async HTTP client (used internally by OpenAI SDK)